[CLEANUP] Move AssertCss logic to Constraint classes
This is the pattern used by PHPUnit, and provides a couple of benefits:
- Constraints can be combined with, e.g., LogicalOr (will be useful for #544);
- The message provided upon a test failure will be clearer.

Also corrected the namespace for `AssertCssTest` for consistency with other
`TestCase` classes.
